Refactor 'value' field to 'settingValue' in Setting class
The name of the 'value' field and its getter and setter methods in the Setting class have been changed to 'settingValue'. This update has been reflected in the corresponding repository tests as well, ensuring that the methods called in the tests match the updated method names.

Change required because update to spring-cloud-starter-parent version (to address some unit test errors) ultimately prevents reserved keyword 'value' being used as a field name.